Support for VPN connection errors
So, I'm not all that proud of this patch, but I had a hard time thinking of a better solution.

While floating IPs and private IPs stay the same, VPN addresses can (and do) change a lot.

On the orchestrator-side, we already have solutions for this in place that allow the orchestrator
to update it's IP address on-the-fly in case your laptop restarts or the connection resets.

But, on the VM-side, if a VPN connection resets, you're SOL without this patch.

The most important part of this patch is not just the VM-side that updates Redis with the
new IP, but really is the SSH-side of this which *dynamically* contacts Redis to get the
most recently-known IP address before each attempt to retry an SSH command.

Patching the retry support was not so straight forward: Now, instead of embedding the IP
address into the SSH command, we lookup that IP on-the-fly each time the ProcessManagement
code needs to repeat the command.

So, it's a little messy, but I tested it quite a bit and it seems stable.

I'm happy to re-work the patch a little bit if it needs any additional cleanups.